+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.73
+
+- Introduces PdfViewerParams.onViewSizeChanged, which is called on view size change
+  - The feature can be used to keep the screen center on device screen rotation (#194)
+
 # 1.0.72
 
 - FIXED: Example code is not compilable
@@ -5,7 +10,7 @@
 - FIXED: Updated podspec file not to download the same archive again and again (#154)
 - Introduces chromium/6555 for all platforms
   - Darwin uses pdfium-apple-v9 (chromium/6555)
-  - Improves memory consumption by pdfium's internal caching feature (#184)
+  - ~~Improves memory consumption by pdfium's internal caching feature (#184)~~
 
 # 1.0.71